requirements.txt
NEURON==8.2.3
numpy==1.26.0
matplotlib==3.8.0
PyYAML==6.0.1